Fix missing acks when under heavy load in gRPC variant
Change queue `offer` to `put` as `put` blocks.

`offer` returns a `boolean` indicating if the insert was successful or
not. Possibly what has been happening in consumers is that under high
load the insert has been unsuccessful, which means the `ack` and `nack`
are discarded entirely. The message was therefore lost (missed acks).