Adds callbacks to HTTP token refreshing on retry and on overall failure
This is an improvement over the existing implementation which only gave
a callback when retries were exhausted overall.

Fixes the refreshing stopping if an exception is raised under the
existing `onTokenRefreshError`, now the docstring clearly states the
user must raise an error in their application. In addition all errors
here are explicitly swallowed.